+++ jupiter-common/src/main/java/org/jupiter/common/atomic/UnsafeAtomicIntegerFieldUpdater.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright (c) 2015 The Jupiter Project
+ *
+ * Licensed under the Apache License, version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jupiter.common.atomic;
+
+import sun.misc.Unsafe;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
+
+/**
+ * jupiter
+ * org.jupiter.common.concurrent.atomic
+ *
+ * Forked from <a href="https://github.com/netty/netty">Netty</a>.
+ */
+final class UnsafeAtomicIntegerFieldUpdater<U> extends AtomicIntegerFieldUpdater<U> {
+    private final long offset;
+    private final Unsafe unsafe;
+
+    UnsafeAtomicIntegerFieldUpdater(Unsafe unsafe, Class<?> tClass, String fieldName) throws NoSuchFieldException {
+        Field field = tClass.getDeclaredField(fieldName);
+        if (!Modifier.isVolatile(field.getModifiers())) {
+            throw new IllegalArgumentException("must be volatile");
+        }
+        if (unsafe == null) {
+            throw new NullPointerException("unsafe");
+        }
+        this.unsafe = unsafe;
+        offset = unsafe.objectFieldOffset(field);
+    }
+
+    @Override
+    public boolean compareAndSet(U obj, int expect, int update) {
+        return unsafe.compareAndSwapInt(obj, offset, expect, update);
+    }
+
+    @Override
+    public boolean weakCompareAndSet(U obj, int expect, int update) {
+        return unsafe.compareAndSwapInt(obj, offset, expect, update);
+    }
+
+    @Override
+    public void set(U obj, int newValue) {
+        unsafe.putIntVolatile(obj, offset, newValue);
+    }
+
+    @Override
+    public void lazySet(U obj, int newValue) {
+        unsafe.putOrderedInt(obj, offset, newValue);
+    }
+
+    @Override
+    public int get(U obj) {
+        return unsafe.getIntVolatile(obj, offset);
+    }
+}